:truck: moving ServiceConfigurationManager into isolable modules in order to allow them to be ignored (overriding not supported)
