+++ service/datastore/internal/src/main/java/org/eclipse/kapua/service/datastore/internal/DatastoreModule.java
@@ -16,22 +16,13 @@ import java.util.Map;
 
 import javax.inject.Singleton;
 
-import org.eclipse.kapua.commons.configuration.CachingServiceConfigRepository;
-import org.eclipse.kapua.commons.configuration.RootUserTester;
-import org.eclipse.kapua.commons.configuration.ServiceConfigImplJpaRepository;
 import org.eclipse.kapua.commons.configuration.ServiceConfigurationManager;
-import org.eclipse.kapua.commons.configuration.ServiceConfigurationManagerCachingWrapper;
-import org.eclipse.kapua.commons.configuration.ServiceConfigurationManagerImpl;
 import org.eclipse.kapua.commons.core.AbstractKapuaModule;
-import org.eclipse.kapua.commons.jpa.EntityCacheFactory;
-import org.eclipse.kapua.commons.jpa.KapuaJpaRepositoryConfiguration;
 import org.eclipse.kapua.commons.jpa.KapuaJpaTxManagerFactory;
 import org.eclipse.kapua.commons.model.domains.Domains;
-import org.eclipse.kapua.commons.util.xml.XmlUtil;
 import org.eclipse.kapua.model.domain.Actions;
 import org.eclipse.kapua.model.domain.Domain;
 import org.eclipse.kapua.model.domain.DomainEntry;
-import org.eclipse.kapua.model.id.KapuaId;
 import org.eclipse.kapua.service.account.AccountService;
 import org.eclipse.kapua.service.authorization.AuthorizationService;
 import org.eclipse.kapua.service.authorization.permission.PermissionFactory;
@@ -48,17 +39,13 @@ import org.eclipse.kapua.service.datastore.internal.converter.ModelContextImpl;
 import org.eclipse.kapua.service.datastore.internal.converter.QueryConverterImpl;
 import org.eclipse.kapua.service.datastore.internal.mediator.DatastoreUtils;
 import org.eclipse.kapua.service.datastore.internal.setting.DatastoreSettings;
-import org.eclipse.kapua.service.datastore.internal.setting.DatastoreSettingsKey;
 import org.eclipse.kapua.service.elasticsearch.client.ElasticsearchClientProvider;
 import org.eclipse.kapua.service.elasticsearch.client.configuration.ElasticsearchClientConfiguration;
 import org.eclipse.kapua.service.elasticsearch.client.rest.MetricsEsClient;
 import org.eclipse.kapua.service.elasticsearch.client.rest.RestElasticsearchClientProvider;
 import org.eclipse.kapua.service.storable.model.id.StorableIdFactory;
-import org.eclipse.kapua.storage.TxContext;
 
 import com.google.inject.Provides;
-import com.google.inject.multibindings.ClassMapKey;
-import com.google.inject.multibindings.ProvidesIntoMap;
 import com.google.inject.multibindings.ProvidesIntoSet;
 
 public class DatastoreModule extends AbstractKapuaModule {
@@ -135,32 +122,5 @@ public class DatastoreModule extends AbstractKapuaModule {
                 messageStoreFacade,
                 metricsDatastore,
                 datastoreSettings);
-    }
-
-    @ProvidesIntoMap
-    @ClassMapKey(MessageStoreService.class)
-    @Singleton
-    ServiceConfigurationManager messageStoreServiceConfigurationManager(
-            RootUserTester rootUserTester,
-            KapuaJpaRepositoryConfiguration jpaRepoConfig,
-            DatastoreSettings datastoreSettings,
-            EntityCacheFactory entityCacheFactory,
-            XmlUtil xmlUtil
-    ) {
-        return new ServiceConfigurationManagerCachingWrapper(new ServiceConfigurationManagerImpl(
-                MessageStoreService.class.getName(),
-                new CachingServiceConfigRepository(
-                        new ServiceConfigImplJpaRepository(jpaRepoConfig),
-                        entityCacheFactory.createCache("AbstractKapuaConfigurableServiceCacheId")
-                ),
-                rootUserTester,
-                xmlUtil
-        ) {
-
-            @Override
-            public boolean isServiceEnabled(TxContext txContext, KapuaId scopeId) {
-                return !datastoreSettings.getBoolean(DatastoreSettingsKey.DISABLE_DATASTORE, false);
-            }
-        });
     }
 }
