+++ service/tag/internal/src/main/java/org/eclipse/kapua/service/tag/internal/TagServiceConfigurationManagerModule.java
@@ -1,0 +1,75 @@
+/*******************************************************************************
+ * Copyright (c) 2021, 2022 Eurotech and/or its affiliates and others
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     Eurotech - initial API and implementation
+ *******************************************************************************/
+package org.eclipse.kapua.service.tag.internal;
+
+import javax.inject.Singleton;
+
+import org.eclipse.kapua.commons.configuration.AccountRelativeFinder;
+import org.eclipse.kapua.commons.configuration.CachingServiceConfigRepository;
+import org.eclipse.kapua.commons.configuration.ResourceLimitedServiceConfigurationManagerImpl;
+import org.eclipse.kapua.commons.configuration.RootUserTester;
+import org.eclipse.kapua.commons.configuration.ServiceConfigImplJpaRepository;
+import org.eclipse.kapua.commons.configuration.ServiceConfigurationManager;
+import org.eclipse.kapua.commons.configuration.ServiceConfigurationManagerCachingWrapper;
+import org.eclipse.kapua.commons.configuration.UsedEntitiesCounterImpl;
+import org.eclipse.kapua.commons.core.AbstractKapuaModule;
+import org.eclipse.kapua.commons.jpa.EntityCacheFactory;
+import org.eclipse.kapua.commons.jpa.KapuaJpaRepositoryConfiguration;
+import org.eclipse.kapua.commons.util.xml.XmlUtil;
+import org.eclipse.kapua.service.tag.TagFactory;
+import org.eclipse.kapua.service.tag.TagRepository;
+import org.eclipse.kapua.service.tag.TagService;
+
+import com.google.inject.Module;
+import com.google.inject.multibindings.ClassMapKey;
+import com.google.inject.multibindings.ProvidesIntoMap;
+
+/**
+ * This module provides the ServiceConfigurationManager for the TagService.
+ * <br><br>
+ * Unfortunately Guice does not support overriding for Map binder entries, therefore if you need to change the behaviour of this ServiceConfigurationManager, just skip this module and define the new
+ * instance in a separate one
+ */
+public class TagServiceConfigurationManagerModule extends AbstractKapuaModule implements Module {
+
+    @Override
+    protected void configureModule() {
+    }
+
+    @ProvidesIntoMap
+    @ClassMapKey(TagService.class)
+    @Singleton
+    ServiceConfigurationManager tagServiceConfigurationManager(
+            TagFactory factory,
+            RootUserTester rootUserTester,
+            AccountRelativeFinder accountRelativeFinder,
+            TagRepository tagRepository,
+            EntityCacheFactory entityCacheFactory,
+            XmlUtil xmlUtil
+    ) {
+        return new ServiceConfigurationManagerCachingWrapper(
+                new ResourceLimitedServiceConfigurationManagerImpl(
+                        TagService.class.getName(),
+                        new CachingServiceConfigRepository(
+                                new ServiceConfigImplJpaRepository(new KapuaJpaRepositoryConfiguration()),
+                                entityCacheFactory.createCache("AbstractKapuaConfigurableServiceCacheId")
+                        ),
+                        rootUserTester,
+                        accountRelativeFinder,
+                        new UsedEntitiesCounterImpl(
+                                factory,
+                                tagRepository
+                        ),
+                        xmlUtil));
+    }
+}
